requirements.txt
# Core ML and Deep Learning
tensorflow==2.11.0
keras-tuner==1.1.0
torch==1.13.0
torchvision==0.14.0
torchaudio==0.13.0

# Data Science and ML
numpy==1.21.4
pandas==1.3.4
scikit-learn==1.1.0
scipy==1.9.0
imbalanced-learn==0.9.0

# Visualization
matplotlib==3.5.1
seaborn==0.11.2
plotly==5.10.0

# Hyperparameter Optimization
optuna==3.0.0
keras-tuner==1.1.0

# Configuration Management
hydra-core==1.2.0
omegaconf==2.2.3

# Experiment Tracking
tensorboard==2.11.0
wandb==0.13.0
mlflow==1.28.0

# Utilities
tqdm==4.64.0
joblib==1.1.0
pyyaml==6.0
click==8.0.3

# Jupyter and Development
jupyter==1.0.0
notebook==6.4.6
ipykernel==6.7.0

# Testing and Code Quality
pytest==7.1.0
pytest-cov==3.0.0
black==22.3.0
flake8==4.0.1
mypy==0.950

# Additional ML Libraries
xgboost==1.6.0
lightgbm==3.3.2
catboost==1.1.1

# Data Processing
openpyxl==3.0.9
xlrd==2.0.1
h5py==3.7.0